cleanup

+++ GUI/Metal/Renderer.swift
@@ -396,13 +396,13 @@ class Renderer: NSObject, MTKViewDelegate {
         let height = (1 - vZoom) * maxHeight
         let bh = ah + vCenter * (maxHeight - height)
                 
-        let texW = CGFloat(EmulatorTexture.width)
-        let texH = CGFloat(EmulatorTexture.height)
+        let texW = EmulatorTexture.width
+        let texH = EmulatorTexture.height
         
-        return CGRect.init(x: CGFloat(bw) / texW,
-                           y: CGFloat(bh) / texH,
-                           width: CGFloat(width) / texW,
-                           height: CGFloat(height) / texH)
+        return CGRect.init(x: CGFloat(bw) / CGFloat(texW),
+                           y: CGFloat(bh) / CGFloat(texH),
+                           width: CGFloat(width) / CGFloat(texW),
+                           height: CGFloat(height) / CGFloat(texH))
     }
     
     // DEPRECATED
